+++ src/ecdh.rs
@@ -29,7 +29,8 @@ pub struct SharedSecret(ffi::SharedSecret);
 impl SharedSecret {
     /// Creates a new shared secret from a pubkey and secret key
     #[inline]
-    pub fn new<C>(secp: &Secp256k1<C>, point: &PublicKey, scalar: &SecretKey) -> SharedSecret {
+    pub fn new(point: &PublicKey, scalar: &SecretKey) -> SharedSecret {
+        let secp = Secp256k1::without_caps();
         unsafe {
             let mut ss = ffi::SharedSecret::blank();
             let res = ffi::secp256k1_ecdh(secp.ctx, &mut ss, point.as_ptr(), scalar.as_ptr());
@@ -102,9 +103,9 @@ mod tests {
         let (sk1, pk1) = s.generate_keypair(&mut thread_rng());
         let (sk2, pk2) = s.generate_keypair(&mut thread_rng());
 
-        let sec1 = SharedSecret::new(&s, &pk1, &sk2);
-        let sec2 = SharedSecret::new(&s, &pk2, &sk1);
-        let sec_odd = SharedSecret::new(&s, &pk1, &sk1);
+        let sec1 = SharedSecret::new(&pk1, &sk2);
+        let sec2 = SharedSecret::new(&pk2, &sk1);
+        let sec_odd = SharedSecret::new(&pk1, &sk1);
         assert_eq!(sec1, sec2);
         assert!(sec_odd != sec2);
     }