+++ src/key.rs
@@ -77,8 +77,8 @@ impl SecretKey {
 
     /// Converts a `SECRET_KEY_SIZE`-byte slice to a secret key
     #[inline]
-    pub fn from_slice<C>(secp: &Secp256k1<C>, data: &[u8])
-                        -> Result<SecretKey, Error> {
+    pub fn from_slice(data: &[u8]) -> Result<SecretKey, Error> {
+        let secp = Secp256k1::without_caps();
         match data.len() {
             constants::SECRET_KEY_SIZE => {
                 let mut ret = [0; constants::SECRET_KEY_SIZE];
@@ -96,8 +96,8 @@ impl SecretKey {
 
     #[inline]
     /// Adds one secret key to another, modulo the curve order
-    pub fn add_assign<C>(&mut self, secp: &Secp256k1<C>, other: &SecretKey)
-                     -> Result<(), Error> {
+    pub fn add_assign(&mut self, other: &SecretKey) -> Result<(), Error> {
+        let secp = Secp256k1::without_caps();
         unsafe {
             if ffi::secp256k1_ec_privkey_tweak_add(secp.ctx, self.as_mut_ptr(), other.as_ptr()) != 1 {
                 Err(InvalidSecretKey)
@@ -109,8 +109,8 @@ impl SecretKey {
 
     #[inline]
     /// Multiplies one secret key by another, modulo the curve order
-    pub fn mul_assign<C>(&mut self, secp: &Secp256k1<C>, other: &SecretKey)
-                     -> Result<(), Error> {
+    pub fn mul_assign(&mut self, other: &SecretKey) -> Result<(), Error> {
+        let secp = Secp256k1::without_caps();
         unsafe {
             if ffi::secp256k1_ec_privkey_tweak_mul(secp.ctx, self.as_mut_ptr(), other.as_ptr()) != 1 {
                 Err(InvalidSecretKey)
@@ -170,8 +170,8 @@ impl PublicKey {
 
     /// Creates a public key directly from a slice
     #[inline]
-    pub fn from_slice<C>(secp: &Secp256k1<C>, data: &[u8])
-                      -> Result<PublicKey, Error> {
+    pub fn from_slice(data: &[u8]) -> Result<PublicKey, Error> {
+        let secp = Secp256k1::without_caps();
 
         let mut pk = unsafe { ffi::PublicKey::blank() };
         unsafe {
@@ -258,7 +258,8 @@ impl PublicKey {
     /// Adds a second key to this one, returning the sum. Returns an error if
     /// the result would be the point at infinity, i.e. we are adding this point
     /// to its own negation
-    pub fn combine<C>(&self, secp: &Secp256k1<C>, other: &PublicKey) -> Result<PublicKey, Error> {
+    pub fn combine(&self, other: &PublicKey) -> Result<PublicKey, Error> {
+        let secp = Secp256k1::without_caps();
         unsafe {
             let mut ret = mem::uninitialized();
             let ptrs = [self.as_ptr(), other.as_ptr()];
@@ -291,9 +292,8 @@ impl<'de> ::serde::Deserialize<'de> for PublicKey {
     fn deserialize<D: ::serde::Deserializer<'de>>(d: D) -> Result<PublicKey, D::Error> {
         use ::serde::de::Error;
 
-        let secp = Secp256k1::without_caps();
         let sl: &[u8] = ::serde::Deserialize::deserialize(d)?;
-        PublicKey::from_slice(&secp, sl).map_err(D::Error::custom)
+        PublicKey::from_slice(sl).map_err(D::Error::custom)
     }
 }
 
@@ -331,24 +331,22 @@ mod test {
 
     #[test]
     fn skey_from_slice() {
-        let s = Secp256k1::new();
-        let sk = SecretKey::from_slice(&s, &[1; 31]);
+        let sk = SecretKey::from_slice(&[1; 31]);
         assert_eq!(sk, Err(InvalidSecretKey));
 
-        let sk = SecretKey::from_slice(&s, &[1; 32]);
+        let sk = SecretKey::from_slice(&[1; 32]);
         assert!(sk.is_ok());
     }
 
     #[test]
     fn pubkey_from_slice() {
-        let s = Secp256k1::new();
-        assert_eq!(PublicKey::from_slice(&s, &[]), Err(InvalidPublicKey));
-        assert_eq!(PublicKey::from_slice(&s, &[1, 2, 3]), Err(InvalidPublicKey));
+        assert_eq!(PublicKey::from_slice(&[]), Err(InvalidPublicKey));
+        assert_eq!(PublicKey::from_slice(&[1, 2, 3]), Err(InvalidPublicKey));
 
-        let uncompressed = PublicKey::from_slice(&s, &[4, 54, 57, 149, 239, 162, 148, 175, 246, 254, 239, 75, 154, 152, 10, 82, 234, 224, 85, 220, 40, 100, 57, 121, 30, 162, 94, 156, 135, 67, 74, 49, 179, 57, 236, 53, 162, 124, 149, 144, 168, 77, 74, 30, 72, 211, 229, 110, 111, 55, 96, 193, 86, 227, 183, 152, 195, 155, 51, 247, 123, 113, 60, 228, 188]);
+        let uncompressed = PublicKey::from_slice(&[4, 54, 57, 149, 239, 162, 148, 175, 246, 254, 239, 75, 154, 152, 10, 82, 234, 224, 85, 220, 40, 100, 57, 121, 30, 162, 94, 156, 135, 67, 74, 49, 179, 57, 236, 53, 162, 124, 149, 144, 168, 77, 74, 30, 72, 211, 229, 110, 111, 55, 96, 193, 86, 227, 183, 152, 195, 155, 51, 247, 123, 113, 60, 228, 188]);
         assert!(uncompressed.is_ok());
 
-        let compressed = PublicKey::from_slice(&s, &[3, 23, 183, 225, 206, 31, 159, 148, 195, 42, 67, 115, 146, 41, 248, 140, 11, 3, 51, 41, 111, 180, 110, 143, 114, 134, 88, 73, 198, 174, 52, 184, 78]);
+        let compressed = PublicKey::from_slice(&[3, 23, 183, 225, 206, 31, 159, 148, 195, 42, 67, 115, 146, 41, 248, 140, 11, 3, 51, 41, 111, 180, 110, 143, 114, 134, 88, 73, 198, 174, 52, 184, 78]);
         assert!(compressed.is_ok());
     }
 
@@ -357,27 +355,24 @@ mod test {
         let s = Secp256k1::new();
 
         let (sk1, pk1) = s.generate_keypair(&mut thread_rng());
-        assert_eq!(SecretKey::from_slice(&s, &sk1[..]), Ok(sk1));
-        assert_eq!(PublicKey::from_slice(&s, &pk1.serialize()[..]), Ok(pk1));
-        assert_eq!(PublicKey::from_slice(&s, &pk1.serialize_uncompressed()[..]), Ok(pk1));
+        assert_eq!(SecretKey::from_slice(&sk1[..]), Ok(sk1));
+        assert_eq!(PublicKey::from_slice(&pk1.serialize()[..]), Ok(pk1));
+        assert_eq!(PublicKey::from_slice(&pk1.serialize_uncompressed()[..]), Ok(pk1));
     }
 
     #[test]
     fn invalid_secret_key() {
-        let s = Secp256k1::new();
         // Zero
-        assert_eq!(SecretKey::from_slice(&s, &[0; 32]), Err(InvalidSecretKey));
+        assert_eq!(SecretKey::from_slice(&[0; 32]), Err(InvalidSecretKey));
         // -1
-        assert_eq!(SecretKey::from_slice(&s, &[0xff; 32]), Err(InvalidSecretKey));
+        assert_eq!(SecretKey::from_slice(&[0xff; 32]), Err(InvalidSecretKey));
         // Top of range
-        assert!(SecretKey::from_slice(&s,
-                                      &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+        assert!(SecretKey::from_slice(&[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
                                         0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFE,
                                         0xBA, 0xAE, 0xDC, 0xE6, 0xAF, 0x48, 0xA0, 0x3B,
                                         0xBF, 0xD2, 0x5E, 0x8C, 0xD0, 0x36, 0x41, 0x40]).is_ok());
         // One past top of range
-        assert!(SecretKey::from_slice(&s,
-                                      &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+        assert!(SecretKey::from_slice(&[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
                                         0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFE,
                                         0xBA, 0xAE, 0xDC, 0xE6, 0xAF, 0x48, 0xA0, 0x3B,
                                         0xBF, 0xD2, 0x5E, 0x8C, 0xD0, 0x36, 0x41, 0x41]).is_err());
@@ -411,22 +406,33 @@ mod test {
 
     #[test]
     fn test_pubkey_from_bad_slice() {
-        let s = Secp256k1::new();
         // Bad sizes
-        assert_eq!(PublicKey::from_slice(&s, &[0; constants::PUBLIC_KEY_SIZE - 1]),
-                   Err(InvalidPublicKey));
-        assert_eq!(PublicKey::from_slice(&s, &[0; constants::PUBLIC_KEY_SIZE + 1]),
-                   Err(InvalidPublicKey));
-        assert_eq!(PublicKey::from_slice(&s, &[0; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE - 1]),
-                   Err(InvalidPublicKey));
-        assert_eq!(PublicKey::from_slice(&s, &[0; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE + 1]),
-                   Err(InvalidPublicKey));
+        assert_eq!(
+            PublicKey::from_slice(&[0; constants::PUBLIC_KEY_SIZE - 1]),
+            Err(InvalidPublicKey)
+        );
+        assert_eq!(
+            PublicKey::from_slice(&[0; constants::PUBLIC_KEY_SIZE + 1]),
+            Err(InvalidPublicKey)
+        );
+        assert_eq!(
+            PublicKey::from_slice(&[0; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE - 1]),
+            Err(InvalidPublicKey)
+        );
+        assert_eq!(
+            PublicKey::from_slice(&[0; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE + 1]),
+            Err(InvalidPublicKey)
+        );
 
         // Bad parse
-        assert_eq!(PublicKey::from_slice(&s, &[0xff; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE]),
-                   Err(InvalidPublicKey));
-        assert_eq!(PublicKey::from_slice(&s, &[0x55; constants::PUBLIC_KEY_SIZE]),
-                   Err(InvalidPublicKey));
+        assert_eq!(
+            PublicKey::from_slice(&[0xff; constants::UNCOMPRESSED_PUBLIC_KEY_SIZE]),
+            Err(InvalidPublicKey)
+        );
+        assert_eq!(
+            PublicKey::from_slice(&[0x55; constants::PUBLIC_KEY_SIZE]),
+            Err(InvalidPublicKey)
+        );
     }
 
     #[test]
@@ -472,12 +478,12 @@ mod test {
         let (mut sk2, mut pk2) = s.generate_keypair(&mut thread_rng());
 
         assert_eq!(PublicKey::from_secret_key(&s, &sk1), pk1);
-        assert!(sk1.add_assign(&s, &sk2).is_ok());
+        assert!(sk1.add_assign(&sk2).is_ok());
         assert!(pk1.add_exp_assign(&s, &sk2).is_ok());
         assert_eq!(PublicKey::from_secret_key(&s, &sk1), pk1);
 
         assert_eq!(PublicKey::from_secret_key(&s, &sk2), pk2);
-        assert!(sk2.add_assign(&s, &sk1).is_ok());
+        assert!(sk2.add_assign(&sk1).is_ok());
         assert!(pk2.add_exp_assign(&s, &sk1).is_ok());
         assert_eq!(PublicKey::from_secret_key(&s, &sk2), pk2);
     }
@@ -490,12 +496,12 @@ mod test {
         let (mut sk2, mut pk2) = s.generate_keypair(&mut thread_rng());
 
         assert_eq!(PublicKey::from_secret_key(&s, &sk1), pk1);
-        assert!(sk1.mul_assign(&s, &sk2).is_ok());
+        assert!(sk1.mul_assign(&sk2).is_ok());
         assert!(pk1.mul_assign(&s, &sk2).is_ok());
         assert_eq!(PublicKey::from_secret_key(&s, &sk1), pk1);
 
         assert_eq!(PublicKey::from_secret_key(&s, &sk2), pk2);
-        assert!(sk2.mul_assign(&s, &sk1).is_ok());
+        assert!(sk2.mul_assign(&sk1).is_ok());
         assert!(pk2.mul_assign(&s, &sk1).is_ok());
         assert_eq!(PublicKey::from_secret_key(&s, &sk2), pk2);
     }
@@ -526,23 +532,19 @@ mod test {
 
     #[test]
     fn pubkey_combine() {
-        let s = Secp256k1::without_caps();
         let compressed1 = PublicKey::from_slice(
-            &s,
             &hex!("0241cc121c419921942add6db6482fb36243faf83317c866d2a28d8c6d7089f7ba"),
         ).unwrap();
         let compressed2 = PublicKey::from_slice(
-            &s,
             &hex!("02e6642fd69bd211f93f7f1f36ca51a26a5290eb2dd1b0d8279a87bb0d480c8443"),
         ).unwrap();
         let exp_sum = PublicKey::from_slice(
-            &s,
             &hex!("0384526253c27c7aef56c7b71a5cd25bebb66dddda437826defc5b2568bde81f07"),
         ).unwrap();
 
-        let sum1 = compressed1.combine(&s, &compressed2);
+        let sum1 = compressed1.combine(&compressed2);
         assert!(sum1.is_ok());
-        let sum2 = compressed2.combine(&s, &compressed1);
+        let sum2 = compressed2.combine(&compressed1);
         assert!(sum2.is_ok());
         assert_eq!(sum1, sum2);
         assert_eq!(sum1.unwrap(), exp_sum);
@@ -550,14 +552,11 @@ mod test {
 
     #[test]
     fn pubkey_equal() {
-        let s = Secp256k1::new();
         let pk1 = PublicKey::from_slice(
-            &s,
             &hex!("0241cc121c419921942add6db6482fb36243faf83317c866d2a28d8c6d7089f7ba"),
         ).unwrap();
         let pk2 = pk1.clone();
         let pk3 = PublicKey::from_slice(
-            &s,
             &hex!("02e6642fd69bd211f93f7f1f36ca51a26a5290eb2dd1b0d8279a87bb0d480c8443"),
         ).unwrap();
 
@@ -591,9 +590,8 @@ mod test {
             0x4a, 0xc8, 0x87, 0xfe, 0x91, 0xdd, 0xd1, 0x66,
         ];
 
-        let s = Secp256k1::new();
-
-        let sk = SecretKey::from_slice(&s, &SK_BYTES).unwrap();
+        let s = Secp256k1::signing_only();
+        let sk = SecretKey::from_slice(&SK_BYTES).unwrap();
         let pk = PublicKey::from_secret_key(&s, &sk);
 
         assert_tokens(&sk, &[Token::BorrowedBytes(&SK_BYTES[..])]);